feat: 데이터베이스에 author 테이블 추가 및 데이터 분리

+++ src/main/java/com/example/schedule/service/ScheduleServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.schedule.service;
 
 import com.example.schedule.dto.*;
 import com.example.schedule.entity.Schedule;
+import com.example.schedule.exception.UnauthorizedAccessException;
 import com.example.schedule.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -13,7 +14,6 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -22,24 +22,23 @@ public class ScheduleServiceImpl implements ScheduleService {
     private final ScheduleRepository scheduleRepository;
 
     @Override
-    public CreateScheduleResponseDto createSchedule(CreateScheduleRequestDto requestDto) {
-        // 일정 생성 로직
+    public void createSchedule(CreateScheduleRequestDto requestDto, Long authorId) {
         // 요청 DTO 를 Entity 로 변환하여 일정 생성
-        Schedule schedule = requestDto.toEntity();
-        Long id = scheduleRepository.saveSchedule(schedule);
-        // 일정 저장 로직 실행 후, 응답 DTO 반한
-        return new CreateScheduleResponseDto(id, schedule.getAuthor(), schedule.getTitle(), schedule.getContent());
+        Schedule schedule = requestDto.toEntity(authorId);
+        scheduleRepository.saveSchedule(schedule);
     }
 
     @Override
-    public List<ScheduleResponseDto> findAllSchedule(String author, String updatedAt) {
+    public List<ScheduleResponseDto> findAllSchedule(String userId, String updatedAt) {
         // 일정 목록 조회 로직
         // 공백으로 입력된 author 나 updatedAt 값을 null 로 변환하여 검색 시 필터링 처리
-        author = convertEmptyToNull(author);
+        userId = convertEmptyToNull(userId);
         updatedAt = convertEmptyToNull(updatedAt);
+
+        Long authorId = scheduleRepository.userIdToAuthorId(userId);
         // 입력된 날짜의 형태 검증
         updatedAt = validateAndFormatDate(updatedAt);
-        return scheduleRepository.findAllSchedule(author, updatedAt);
+        return scheduleRepository.findAllSchedule(authorId, updatedAt);
     }
 
     @Override
@@ -51,18 +50,15 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public ScheduleResponseDto updateSchedule(Long id, UpdateScheduleRequestDto requestDto) {
-        // 일정 수정 로직
-        // author, title, content 값이 있을 경우 각각 수정
+    public void updateSchedule(Long id, UpdateScheduleRequestDto requestDto, Long sessionAuthorId) {
 
-        // id 유효성 검사
-        validateId(id);
-        // password 유효성 검사
-        validatePassword(id, requestDto.getPassword());
-        // author 값 입력 시 수정 로직
-        if(requestDto.getAuthor() != null){
-            updateAuthor(id, requestDto.getAuthor());
+        if (isAuthorized(id, sessionAuthorId)) {
+            throw new UnauthorizedAccessException("수정 권한이 없습니다.");
         }
+
+        validateId(id);
+
+        // author 값 입력 시 수정 로직
         // title 값 입력 시 수정 로직
         if(requestDto.getTitle() != null){
             updateTitle(id, requestDto.getTitle());
@@ -71,46 +67,28 @@ public class ScheduleServiceImpl implements ScheduleService {
         if(requestDto.getContent() != null){
             updateContent(id, requestDto.getContent());
         }
-        // 수정된 데이터 Dto 로 받아오기
-        Optional<ScheduleResponseDto> updatedSchedule = scheduleRepository.findScheduleById(id);
-
-        return updatedSchedule
-                .orElseThrow(() -> new NoSuchElementException("일치하는 일정이 없습니다."));
     }
 
     @Override
-    public void deleteSchedule(Long id, DeleteScheduleRequestDto requestDto) {
+    public void deleteSchedule(Long id, Long sessionAuthorId) {
+
+        if (isAuthorized(id, sessionAuthorId)) {
+            throw new UnauthorizedAccessException("수정 권한이 없습니다.");
+        }
         // ID에 해당하는 일정 삭제
         // 유효성 검사 후 삭제 수행
-
         // 주어진 ID가 존재하는지 확인
         // ID가 없으면 NoSuchElementException 발생
         validateId(id);
 
-        // 주어진 ID와 비밀번호가 일치하는지 확인
-        // 비밀번호가 일치하지 않으면 IllegalArgumentException 발생
-        validatePassword(id, requestDto.getPassword());
-
         // 일정 삭제 로직 실행
         scheduleRepository.deleteSchedule(id);
-
     }
 
     private void validateId(Long id) {
         if (!scheduleRepository.isValidId(id)) {
-            throw new NoSuchElementException("일치하는 ID가 없습니다.");
+            throw new NoSuchElementException(id+"번 게시물이 없습니다.");
         }
-    }
-
-    private void validatePassword(Long id, String password) {
-        if (!scheduleRepository.isValidPassword(id, password)) {
-            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
-        }
-    }
-
-    private void updateAuthor(Long id, String author){
-        // 작성자 수정
-        scheduleRepository.updateAuthor(id, author);
     }
 
     private void updateTitle(Long id, String title){
@@ -141,5 +119,13 @@ public class ScheduleServiceImpl implements ScheduleService {
             }
         }
         return null;
+    }
+
+    private boolean isAuthorized(Long id, Long sessionAuthorId) {
+        // 게시물 번호의 author id 조회하기
+        Long authorId = scheduleRepository.idToAuthorId(id);  // DB 조회
+
+        // 로그인한 사용자와 일정 작성자 비교
+        return !authorId.equals(sessionAuthorId);
     }
 }
